Add weather forecast API documentation and handle exceptions in the backend

+++ backend/src/main/java/com/pweb/backend/services/UserService.java
@@ -8,9 +8,10 @@ import com.pweb.backend.dao.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -30,6 +31,10 @@ public class UserService {
 
 
     public User registerUser(AuthenticationController.RegisterRequest registerRequest) {
+        // check if username is already taken
+        if (userRepository.existsByUsername(registerRequest.username)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Username already taken");
+        }
         User user = new User();
 
         user.setUsername(registerRequest.username);
@@ -49,8 +54,11 @@ public class UserService {
         return userRepository.findAll().stream().map(User::getUsername).toList();
     }
 
-//    @Transactional
+    //    @Transactional
     public void deleteUser(Integer id) {
+        if (!userRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found");
+        }
         userRepository.deleteById(id);
     }
